backend/requirements.txt
fastapi
sqlmodel
aiosqlite
asyncpg
pymysql
aiomysql
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic-settings
bcrypt==4.0.1